WARP testnet smoke-test + dokumentace a drobné opravy.
- Stellar testnet execute_mint PASS (0.001 ZION, tx 2cbe550f...8d34e).
- Solana/Bitcoin polling OK, live mint/spend zatím bez funds.
- Aktualizována dokumentace: CONTRACT_ADDRESSES.md, DEPLOY.md, StatusV3.md.
- Přidána kořenová zpráva WARP_ZPRAVA_2026-08-06.md.
- Opravy kompilace zion-pool (SocketAddr, reward) a clippy warning ve WARP executoru.
- Všechny workspace testy a clippy procházejí.

+++ V31/L2/multichain/src/warp/executor.rs
@@ -186,7 +186,7 @@ impl OutboundExecutor {
         .ok_or(WarpError::DecimalOverflow {
             from_decimals: 6,
             to_decimals: transfer.dest_chain.decimals,
-        })? as u128;
+        })?;
 
         // Build mint instruction
         let mint_instruction = MintInstruction {
